ref(replay): Centralize ReplayTable cell rendering into replayTableColumns
